+++ lib/src/utils/utils.dart
@@ -6,17 +6,28 @@ class Utils {
   String get generateId =>
       'BabakCode${random.nextInt(99999)}${random.nextInt(99999)}';
 
-  String relatedJs({required String id}) => """
-cameraOrbit$id = (a, b, c) => {document.querySelector('#$id').cameraOrbit = `\${a}deg \${b}deg \${c}m`}
-cameraTarget$id = (x, y, z) => {document.querySelector('#$id').cameraTarget = `\${x}m \${y}m \${z}m`}
-customEvaluate$id = (code) => { eval(code) }
-animationName$id = (set) => { document.querySelector('#$id').setAttribute('animation-name', set); }
-availableAnimations$id = (set) => { return document.querySelector('#$id').availableAnimations; }
-autoRotate$id = (set) => { document.querySelector('#$id').setAttribute('auto-rotate', set ?? false); }
-autoPlay$id = (set) => { (set ?? false) ? document.querySelector('#$id').play() : document.querySelector('#$id').pause(); }
-cameraControls$id = (set) => { document.querySelector('#$id').setAttribute('camera-controls', set ?? false); }
-variantName$id = (set) => { document.querySelector('#$id').variantName = set; }
-play$id = ({repetitions}) => { document.querySelector('#$id').play({repetitions}); }
-pause$id = () => { document.querySelector('#$id').pause(); }
-""";
+  String relatedJs({required String id}) => '''
+  const o3d$id = document.querySelector('#$id');
+  const defaultCameraOrbit$id = o3d$id.cameraOrbit;
+customJsCode$id = (code) => { eval(code) }
+customEvaluateWithResult$id = (code) => { return eval(code) }
+  ''';
+
+  /// Covert tuple string to list
+  /// [tupleToList]
+  static String tupleToList(String input) {
+    String trimmedString = input.trim();
+    if (trimmedString.isNotEmpty && trimmedString.length > 2) {
+      if (trimmedString[0] == '(') {
+        trimmedString = trimmedString.replaceFirst('(', '[');
+      }
+      if (trimmedString[trimmedString.length - 1] == ')') {
+        trimmedString =
+            '${trimmedString.substring(0, trimmedString.length - 2)}]';
+      }
+      return trimmedString;
+    } else {
+      return '[]';
+    }
+  }
 }